Bug 1475306 - Changes to the detection of extensions already migrated to ExtensionStorageIDB and rename successfully migrated JSONFile. r=mixedpuppy
This patch applies the following changes to the storage.local data migration behaviors:

- An about:config preference is set when an extension has been migrated successfully to the
  storage.local IndexedDB backend (cleared automatically if the addon is uninstalled).
- If the above about:config preference is set, the storage.local IndexedDB backend is enabled
  without attempting to open an IndexedDB connection for the new backend.
- While migrating an extension, if we fail to open the IndexedDB connection, the data migration
  is cancelled and the storage.local API is going to fallback to the storage.local JSONFile backend
  (until the next extension startup, when a new data migration is going to be tried).
- When a migration is completed successfully, the old JSONFile is renamed (by appending ".migrated"
  to its original file name) instead of being removed.

MozReview-Commit-ID: LPM0fQUagTd

UltraBlame original commit: 1fafb8e07051ccf79a2fa5364d334baa7ac83f10

+++ toolkit/components/extensions/ExtensionStorage.jsm
@@ -137,8 +137,12 @@ var ExtensionStorage = {
 
 
 
-  clearCachedFile(extensionId) {
-    this.jsonFilePromises.delete(extensionId);
+  async clearCachedFile(extensionId) {
+    let promise = this.jsonFilePromises.get(extensionId);
+    if (promise) {
+      this.jsonFilePromises.delete(extensionId);
+      await promise.then(jsonFile => jsonFile.finalize());
+    }
   },
